ms-download/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
cassandra-driver==3.29.1
minio==7.2.7
python-jose[cryptography]==3.3.0
httpx==0.27.0
